Fix annotate sim for z/OS.
For the version of perl on z/OS we must explicitly create Math::BigInt
instead of using hex.
Address may use upper case and may not start with 0x.

Annotates with stub names now but there is still a bug with invoke
depth because JSEntryStub+0 is not annotated correctly. It looks like
this is caused by a bug in the output of print-all-code itself for the
first instruction of JSEntryStub.